+++ repo/src/main/java/com/app/repo/impl/SalesStandRepositoryImpl.java
@@ -5,11 +5,7 @@ package com.app.repo.impl;
 import com.app.model.SalesStand;
 import com.app.repo.generic.AbstractCrudRepository;
 
-public class SalesStandRepositoryImpl extends AbstractCrudRepository<SalesStand, Integer> {
-
-
-
-
+public class SalesStandRepositoryImpl extends AbstractCrudRepository<SalesStand, Long> {
